Refactor api client
